modul accounting

+++ app/Http/Controllers/PermissionController.php
@@ -10,6 +10,10 @@ class PermissionController extends Controller
 {
     public function index(Request $request)
     {
+        if (!auth()->user()->can('manage permission')) {
+            abort(403, 'Anda tidak memiliki izin akses.');
+        }
+
         $query = Permission::query();
 
         if ($request->has('search')) {
@@ -23,6 +27,10 @@ class PermissionController extends Controller
 
     public function create()
     {
+        if (!auth()->user()->can('create permission')) {
+            abort(403, 'Anda tidak memiliki izin untuk menambah user.');
+        }
+
         return view('permission.create');
     }
 
@@ -33,6 +41,7 @@ class PermissionController extends Controller
             'group' => 'nullable|string',
         ]);
 
+        /* dd($request); */
         Permission::create([
             'name' => $request->name,
             'group' => $request->group,
@@ -44,6 +53,10 @@ class PermissionController extends Controller
 
     public function edit($id)
     {
+        if (!auth()->user()->can('edit permission')) {
+            abort(403, 'Anda tidak memiliki izin untuk menambah user.');
+        }
+
         $permission = Permission::findOrFail($id);
         return view('permission.edit', compact('permission'));
     }
@@ -64,6 +77,10 @@ class PermissionController extends Controller
 
     public function destroy($id)
     {
+        if (!auth()->user()->can('delete permission')) {
+            abort(403, 'Anda tidak memiliki izin untuk menambah user.');
+        }
+        
         $permission = Permission::findOrFail($id);
         $permission->delete();
